refactor: move FindPlantingsParameters
