+++ backend/src/service/plantings.rs
@@ -11,7 +11,8 @@ use crate::model::dto::plantings::{
     NewPlantingDto, PlantingDto, PlantingSearchParameters, UpdatePlantingDto,
 };
 use crate::model::dto::TimelinePage;
-use crate::model::entity::plantings::{FindPlantingsParameters, Planting};
+use crate::model::entity::plantings::Planting;
+use crate::model::entity::plantings_impl::FindPlantingsParameters;
 
 /// Time offset in days for loading plantings in the timeline.
 const TIME_LINE_LOADING_OFFSET_DAYS: u64 = 356;
